Added a new resource
Hi, I have added a new resource for Python dictionaries that I believe will add some value to this for loop post. Got this tutorial online while looking for some resources to learn python. Topics were explained in detail and found it to be useful and thought of adding it here. Hope this small contribution will help other learners.